+++ cmd/hidden_lake/messenger/internal/handler/friends_chat.go
@@ -62,7 +62,7 @@ func FriendsChatPage(pStateManager state.IStateManager, msgLimit uint64) http.Ha
 			return
 		}
 
-		client := pStateManager.GetClient().Service()
+		client := pStateManager.GetClient()
 		myPubKey, err := client.GetPubKey()
 		if err != nil {
 			fmt.Fprint(pW, errors.WrapError(err, "error: read public key"))