+++ cmd/hidden_lake/messenger/pkg/app/state/types.go
@@ -5,7 +5,6 @@ import (
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/database"
 	"github.com/number571/go-peer/cmd/hidden_lake/messenger/internal/utils"
 	hls_client "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/client"
-	hlt_client "github.com/number571/go-peer/cmd/hidden_lake/traffic/pkg/client"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 )
 
@@ -16,7 +15,7 @@ type STemplateState struct {
 
 type SStorageState struct {
 	FPrivKey     string            `json:"priv_key"`
-	FConnections []string          `json:"connections"`
+	FConnections []*SConnection    `json:"connections"`
 	FFriends     map[string]string `json:"friends"`
 }
 
@@ -28,18 +27,19 @@ type IStateManager interface {
 	OpenState([]byte) error
 	CloseState() error
 
-	GetClient() IClient
+	GetClient() hls_client.IClient
 	GetWrapperDB() database.IWrapperDB
 	GetTemplate() *STemplateState
 
 	AddFriend(string, asymmetric.IPubKey) error
 	DelFriend(string) error
 
-	AddConnection(string) error
+	GetConnections() ([]IConnection, error)
+	AddConnection(string, bool) error
 	DelConnection(string) error
 }
 
-type IClient interface {
-	Service() hls_client.IClient
-	Traffic() hlt_client.IClient
+type IConnection interface {
+	GetAddress() string
+	IsBackup() bool
 }